requirements.txt
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.3
python-telegram-bot==20.7
python-dotenv==1.0.0
ta==0.11.0
streamlit==1.31.1
plotly==5.18.0 